Refactored account module.
