+++ lib/helpers.js
@@ -6,6 +6,7 @@
 
 const rlp          = require('rlp');
 const keccak256    = require('keccak256');
+const MerkleTree   = require('merkletreejs');
 const ethUtil      = require('ethereumjs-util');
 const ethereumAbi  = require('ethereumjs-abi');
 const ethereumTx   = require('ethereumjs-tx');
@@ -77,4 +78,36 @@ exports.toTxObject = function toTxObject(serializedTx) {
     tx.from = '0x' + keccak256(publicKey).slice(12).toString('hex');
 
     return tx;
+};
+
+/**
+ * Freshly created account.
+ *
+ * @param  {String} address     Name which links thee to this world of blockchain.
+ * @param  {Number} [balance=0] Initial balance (positive in case of genesis allocation).
+ * @return {Object}             Empty account.
+ */
+exports.emptyAccount = function emptyAccount(address, balance=0) {
+    return {
+        address,
+        balance,
+        locked: 0,
+        nonce:  0,
+        rating: 0,
+        certificates: [],
+        votes:       []
+    };
+};
+
+/**
+ * Generate merkle tree root.
+ *
+ * @param  {String[]} rawLeaves
+ * @return {String}             Tree root.
+ */
+exports.merkleRoot = function merkleRoot(rawLeaves) {
+    const leaves = rawLeaves.map(rawLeaf => keccak256(rawLeaf));
+    const tree   = new MerkleTree(leaves, keccak256);
+
+    return '0x' + tree.getRoot().toString('hex');
 };
