+++ lib/types.js
@@ -11,7 +11,7 @@ exports.block = {
     txRoot:       'string',
     receiptsRoot: 'string',
     state:        'object[]',
-    tx:           'object[]',
+    transactions: 'object[]',
     receipts:     'object[]'
 };
